\name{hiv}
\alias{hiv}
\docType{data}
\title{
HIV Data Set
}
\description{
These data contain effect size estimates and standard errors obtained from 
gene expression measurements on 7680 genes across 8 samples.
}
\usage{data(hiv)}
\format{
  A data frame with 7680 observations on the following 2 variables.
  \describe{
    \item{\code{estimates}}{a vector of effect size estimates}
    \item{\code{std.err}}{standard errors associated with effect size estimates}
  }
}
\references{
van't Wout, et. al. (2003), Cellular gene 
expression upon human immuno-deficiency virus type 1 infection of
CD4+-T-Cell lines, \emph{Journal ofVirology}, \bold{77}, 1392--1402.
}
\keyword{datasets}
