\name{MCMCtest}
\alias{MCMCtest}
\docType{data}
\title{
MCMC sample output
}
\description{
a matrix of test MCMC output
}
\usage{data(MCMCtest)}
\format{
A 2000 x 400 numeric matrix. Data in the ith row should be thought of 
as a sample from the posterior for the ith case of interest.
}
\seealso{
\code{\link{rvaluesMCMC}}
}
\keyword{datasets}
