\name{NBA1314}
\alias{NBA1314}
\docType{data}
\title{
National Basketball Association, free throw data, 2013-2014 season
}
\description{Free throw statistics on 482 active players, 2013-2014 season}
\usage{data(NBA1314)}
\format{
  A data frame with 482 players (rows) variables including.
  \describe{
    \item{\code{PLAYER}}{name of player} 
    \item{\code{TEAM}}{player's team}
    \item{\code{GP}}{games played}
    \item{\code{PPG}}{points per game} }
    \item{\code{FTA}}{free throws attempted} }
    \item{\code{FTM}}{free throws made} }
    \item{\code{FTprop}}{FTA/FTM} }
  }
}
\details{
Data obtained from ESPN.
}
\references{See data analyzed in Henderson and Newton, 2015}
}
\keyword{datasets}
